tests: rework tests after changes

+++ tests/workflow/Snakemake_setup.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import shutil
 import stat
@@ -9,16 +10,22 @@ from omnibenchmark.workflow.snakemake import SnakemakeEngine
 TO_CLEANUP = [".snakemake", "out", "Snakefile", "snakemake.log"]
 
 
-class SnakemakeSetup:
-    def __init__(self, benchmark_file: Path) -> None:
-        self.benchmark_file = benchmark_file
+def raise_if_file_not_found(file_path: str):
+    if not os.path.exists(file_path):
+        raise FileNotFoundError(f"File {file_path} does not exist.")
 
-        if os.path.exists(benchmark_file):
-            self.benchmark = Benchmark(benchmark_file)
-            self.workflow = SnakemakeEngine()
-            self._print_benchmark()
-        else:
-            raise FileNotFoundError(f"Benchmark file {benchmark_file} does not exist.")
+
+class SnakemakeSetup:
+    def __init__(
+        self, benchmark_file: Path, keep_files: bool = False, cwd: str = ""
+    ) -> None:
+        raise_if_file_not_found(benchmark_file.as_posix())
+        self.benchmark_file = benchmark_file
+        self.benchmark = Benchmark(benchmark_file)
+        self.workflow = SnakemakeEngine()
+        self.keep_files = keep_files
+        self.cwd = cwd
+        self._print_benchmark()
 
     def __enter__(self):
         return self
@@ -27,9 +34,11 @@ class SnakemakeSetup:
         self._cleanup_snakemake()
 
     def _cleanup_snakemake(self):
-        current_dir = os.getcwd()
+        if self.keep_files:
+            logging.getLogger().warning("File cleanup skipped")
+            return
         for file in TO_CLEANUP:
-            file_path = os.path.join(current_dir, file)
+            file_path = os.path.join(self.cwd, file)
             if os.path.exists(file_path):
                 if os.path.isfile(file_path):
                     os.remove(file_path)
